Offer alternative way to build in OpenShift
See https://github.com/opendevstack/ods-jenkins-shared-library/issues/3.

Instead of moving the artifacts to Docker via Nexus, we upload the
artifact directly. This should be a bit faster (as downloading is not
necessary), it should be less brittle (as paths in Jenkins and
Dockerfile do not need to match), requires less params (e.g. Nexus
credentials) and less code (as archiving is done automatically by OC).

Also, we now follow the build instead of simply waiting for it, removing
the need to go into OpenShift to see what happened with the build.