Deploy via Tailor
A deployment might be more than just updating an image, e.g. it might also
involve adding an environment variable, or changing constraints etc. Because of
this, a deployment should be a set of of JSON patchsets, not just an image tag update.

When an "openshift" folder is present, we use the "tailor" binary in the
jenkins-slave-base image to apply templates in that folder. The behaviour is customizable by
adding a "Tailorfile" (or a more specific, per-environment version of it).

The actual rollout is triggered explicitly (or watched if triggered
implicitly), which also allows us to monitor the progress and fail the build if
the rollout does not work.