Fix BitBucket HTTP calls when user credentials contain dollar signs
When credentials contain dollar signs, Jenkins does not automatically
escape them. When the credentials is used in the shell then, the
dollar sign denotes the start of a variable. Most likely, that variable
will not exist so it expands to an empty string. Therefore, we escape
dollar signs so that they are not expanded, but taken literally.

See also https://issues.jenkins-ci.org/browse/JENKINS-40734/.