Add nexusUrl in favour of nexusHost
nexusHost contained the value of NEXUS_HOST previously, which, despite
its name, contained not just the host but also the scheme part.

For consistency, we now prefer NEXUS_URL, and expose this as nexusUrl.
nexusHost will now NOT include the scheme anymore. This is a breaking
change. Users of ODS will be affected and need to update their
Jenkinsfiles accordingly. There is no way to make this non-breaking, we
could only delay the impact by setting NEXUS_HOST/nexusHost to the same
value as nexusUrl, but that would be weird end end up being a breaking
change when we adapt it to be host only in the future.

Addresses part of #74.