Convert Snyk stage to new style
This commit also adapts the behaviour: previously the stage would only
fail if `failOnSnykScanVulnerabilities` was `true` and `snyk test` failed.

Now the stage fails on every error within the stage, such as `snyk auth`
failing, regardless if `failOnSnykScanVulnerabilities` is `true` or not.

Setting `failOnSnykScanVulnerabilities` to `false` only prevents an
error when `snyk test` fails.

Further, the stage now allows to override the `--project-name` flag
(which is defaulted to `componentId`), and it adds more defaults
(`projectId` for `organisation` and `build.gradle` for `buildFile`).