Perform PR analysis
When a branch is eligible to be scanned, and it is not one of the
long-lived branches, we ask SonarQube to perform PR analysis and
decorate the associated pull request in Bitbucket.

As a side-effect, we now have a Bitbucket token for the cd_user. This is
required as PR decoration is token based. However, a token is good to
have anyway, and we can move other Bitbucket API calls over to the
BitbucketService in the future, making use of the token instead of the
password.

Closes #174.