Do not fail when secret cannot be created
We might not be able to create it because it has been created already by
another job at the same time. This is super unlikely to happen in real
life, but does happen during parallel tests.

Fixes #432.

The script errors later anyway if the Jenkins credential cannot be
found (see
https://github.com/opendevstack/ods-jenkins-shared-library/blob/e63d0976d42aad08fddfebfbd8fbdbad34effbf0/src/org/ods/services/BitbucketService.groovy#L243).