Prefer issue ID in commit message
The Jira issue ID can be given in either branch name or commit message.
Putting the ID into the commit message has the advantage that it is
still accessible when the commit is accessed from a different branch,
and that it is possible that every commit is connected to a Jira ticket.