Support Deployment resources in component pipeline
Closes #295.

Also deals with multiple pods (replicas) in the OpenShiftService, laying
some groundwork for #390.

Note that due to `rollout restart` being introduced in Kubernetes 1.15,
Deployment resources are not supported in OpenShift 3.11 at the moment.
This could change in future PRs if we change how we trigger rollouts
(specifically, by not using the `latest` tag but instead update the pod
template to different image tags for every rollout). Further, use of
`Deployment` resources is not (yet) possible when the orchestration
pipeline is being used.

Also note that the introduction of the config option `deploymentKind` in
`odsComponentStageRolloutOpenShiftDeployment` will likely be temporary.
https://github.com/opendevstack/ods-jenkins-shared-library/issues/496
will require to refactor the stage so that it can deal with the whole
component, not individual Deployment or DeploymentConfig resources.