Remove automatic versioned dev env creation
The option to use a versioned env still exists, but the creation has to
be manual now. This cicrumvents issues with `self-provisioner`
permissions and missing `admin` permissions of the `jenkins`
serviceaccount.

Going forward, this allows us to build images in the `*-cd` project,
which is easier to manage, and frees us from requiring the `*-dev` env
to be on the same cluster. This can be handy if e.g. you want to have
build infrastructure on a non-internet facing cluster, but deploy into
an internet-facing cluster.

This change is related to
https://github.com/opendevstack/ods-jenkins-shared-library/pull/512.