Always build image in orchestration pipeline
Before
https://github.com/opendevstack/ods-jenkins-shared-library/pull/517, an
image was built if an environment was determined. In an orchestration
context, an environment is always given, which means an image would have
been built always. #517 accidentally changed that behaviour - an image
would not have been built when no environment was selected via the
branchToEnvironmentMapping. This could have caused issue under certain
configuration combinations.