Move repo execution from util to each stage
This should make it much easier to follow what's going on. The previous
version was so difficult to understand that each time I looked at it, I
had to trace the flow again. This was even more complicated by the fact
that the previous version used incorrect method types (which Jenkins
ignores, and no tests were written for those methods).

The new version should be simpler to understand, and has tests. Further, all
stages are now type checked, and there is at least one test run to
verify the bare minimum.